orders/src/routes/delete.ts
import {
  NotAuthError,
  NotFoundError,
  OrderStatus,
  requireAuth,
} from "@ticketscourse/common";
import express, { Request, Response } from "express";
import { Order } from "../models/order";
import { OrderCancelledPublisher } from "../events/publishers/order-cancelled-publisher";
import { natsWrapper } from "../nats-wrapper";

const router = express.Router();

router.delete(
  "/api/orders/:orderId",
  requireAuth,
  async (req: Request, res: Response) => {
    const order = await Order.findById(req.params.orderId).populate("ticket");
    //populate => send the ticket with the data

    if (!order) {
      throw new NotFoundError();
    }

    if (order.userId !== req.currentUser!.id) {
      throw new NotAuthError();
    }

    order.status = OrderStatus.Canceled;
    await order.save();

    // publish even say we cancelled the order
    new OrderCancelledPublisher(natsWrapper.client).publish({
      id: order.id,
      version: order.version,
      ticket: {
        id: order.ticket.id,
      },
    });

    res.status(201).send(order);
  }
);

export { router as deleteOrderRouter };
